python-server/requirements.txt
fastapi
uvicorn
langchain==0.1.14
crewai==0.22.6
google-generativeai
python-dotenv
sentence-transformers